PC is now +2'd after instruction fetch. +4's to PC have been changed to +2's as they add up.  Fixed an error in Util that was completely breaking sprite drawing.
Sprite drawing is done by loading the sprite's location from register I, our instruction that sets register I (Annn) was using Util.GetBits12to0, which returns a ushort, but was getting cast to a byte before returning! Changing this simple bit has *mostly* fixed sprite drawing.

+++ CSIP8/CSIP8/Util.cs
@@ -45,7 +45,7 @@ namespace CSIP8
 
         public static ushort GetBits12To0(ushort us)
         {
-            return (byte)(us & 0x0FFF);
+            return (ushort)(us & 0x0FFF);
         }
 
         /* Quarters */
